Entities commands data validation moved from validators to command handlers, tests fixed

+++ backend/Library.Application/AuthorUseCases/Commands/DeleteAuthorHandler.cs
@@ -1,5 +1,8 @@
+using Library.Domain.Abstractions;
+using Library.Domain.Entities;
 using Library.Domain.Specifications;
 using Library.Domain.Specifications.AuthorSpecification;
+using Library.Domain.Specifications.BookSpecifications;
 
 namespace Library.Application.AuthorUseCases.Commands;
 
@@ -14,9 +17,17 @@ public class DeleteAuthorHandler : IRequestHandler<DeleteAuthorCommand, Unit>
 
     public async Task<Unit> Handle(DeleteAuthorCommand request, CancellationToken cancellationToken)
     {
-        var spec = new AuthorByIdSpecification(request.Id);
-        var author = await _unitOfWork.AuthorRepository.FirstOrDefault(spec, cancellationToken);
+        var authorSpec = new AuthorByIdSpecification(request.Id);
+        var bookSpec = new BookCatalogSpecification(null, null, request.Id, null);
 
+        var author = await _unitOfWork.AuthorRepository.FirstOrDefault(authorSpec, cancellationToken);
+        if (author == null)
+            throw new ValidationException($"Author with specified ID ({request.Id}) doesn't exist");
+
+        if (await _unitOfWork.BookRepository.CountAsync(bookSpec, cancellationToken) != 0)
+            throw new ValidationException($"Author being deleted must have no books");
+
+        
         _unitOfWork.AuthorRepository.Delete(author);
         await _unitOfWork.SaveChangesAsync();
